Remove copying of block_id from amendment creation
