+++ tests/system/action/motion/test_create_amendment.py
@@ -1,12 +1,13 @@
+from openslides_backend.permissions.permissions import Permissions
 from tests.system.action.base import BaseActionTestCase
 
 
 class MotionCreateAmendmentActionTest(BaseActionTestCase):
     def setUp(self) -> None:
         super().setUp()
-        # create parent motion and workflow
         self.set_models(
             {
+                "meeting/222": {"is_active_in_organization_id": 1},
                 "motion_workflow/12": {
                     "name": "name_workflow1",
                     "first_state_id": 34,
@@ -18,16 +19,12 @@ class MotionCreateAmendmentActionTest(BaseActionTestCase):
                     "sort_child_ids": [],
                     "meeting_id": 222,
                 },
+                "motion_category/12": {"meeting_id": 222},
+                "motion_block/13": {"meeting_id": 222},
             }
         )
 
     def test_create_amendment(self) -> None:
-        self.set_models(
-            {
-                "meeting/222": {"is_active_in_organization_id": 1},
-                "user/1": {"meeting_ids": [222]},
-            }
-        )
         response = self.request(
             "motion.create",
             {
@@ -36,14 +33,37 @@ class MotionCreateAmendmentActionTest(BaseActionTestCase):
                 "workflow_id": 12,
                 "lead_motion_id": 1,
                 "text": "text_test1",
+                "block_id": 13,
+                "category_id": 12,
             },
         )
         self.assert_status_code(response, 200)
-        model = self.get_model("motion/2")
-        assert model.get("title") == "test_Xcdfgee"
-        assert model.get("meeting_id") == 222
-        assert model.get("lead_motion_id") == 1
-        assert model.get("text") == "text_test1"
+        self.assert_model_exists(
+            "motion/2",
+            {
+                "meeting_id": 222,
+                "lead_motion_id": 1,
+                "title": "test_Xcdfgee",
+                "text": "text_test1",
+                "block_id": 13,
+                "category_id": 12,
+            },
+        )
+
+    def test_create_amendment_inherited_category(self) -> None:
+        self.set_models({"motion/1": {"category_id": 12, "block_id": 13}})
+        response = self.request(
+            "motion.create",
+            {
+                "title": "test_Xcdfgee",
+                "meeting_id": 222,
+                "lead_motion_id": 1,
+                "workflow_id": 12,
+                "text": "text",
+            },
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("motion/2", {"block_id": None, "category_id": 12})
 
     def test_create_amendment_default_workflow(self) -> None:
         self.set_models(
@@ -270,8 +290,134 @@ class MotionCreateAmendmentActionTest(BaseActionTestCase):
                 "title": "test_Xcdfgee",
                 "meeting_id": 222,
                 "workflow_id": 12,
+                "lead_motion_id": 1,
                 "text": "text",
             },
         )
         self.assert_status_code(response, 400)
         assert "Reason is required" in response.json["message"]
+
+    def test_create_amendment_permission(self) -> None:
+        self.create_meeting(222)
+        self.user_id = self.create_user("user")
+        self.login(self.user_id)
+        self.set_user_groups(self.user_id, [224])
+        self.set_group_permissions(
+            224,
+            [Permissions.Motion.CAN_CREATE, Permissions.Motion.CAN_CREATE_AMENDMENTS],
+        )
+        response = self.request(
+            "motion.create",
+            {
+                "title": "test_Xcdfgee",
+                "meeting_id": 222,
+                "workflow_id": 12,
+                "text": "test",
+                "lead_motion_id": 1,
+            },
+        )
+        self.assert_status_code(response, 200)
+
+    def test_create_amendment_no_permission(self) -> None:
+        self.create_meeting(222)
+        self.user_id = self.create_user("user")
+        self.login(self.user_id)
+        self.set_user_groups(self.user_id, [224])
+        self.set_group_permissions(224, [Permissions.Motion.CAN_CREATE])
+        response = self.request(
+            "motion.create",
+            {
+                "title": "test_Xcdfgee",
+                "meeting_id": 222,
+                "workflow_id": 12,
+                "text": "test",
+                "lead_motion_id": 1,
+            },
+        )
+        self.assert_status_code(response, 403)
+        assert (
+            "Missing Permission: motion.can_create_amendments"
+            in response.json["message"]
+        )
+
+    def test_create_amendment_non_admin(self) -> None:
+        self.create_meeting(222)
+        self.user_id = self.create_user("user")
+        self.login(self.user_id)
+        self.set_user_groups(self.user_id, [224])
+        self.set_group_permissions(
+            224,
+            [
+                Permissions.Motion.CAN_CREATE,
+                Permissions.Motion.CAN_CREATE_AMENDMENTS,
+                Permissions.Motion.CAN_MANAGE,
+            ],
+        )
+        self.set_models(
+            {
+                "motion/1": {
+                    "category_id": 12,
+                    "block_id": 13,
+                },
+            }
+        )
+        response = self.request(
+            "motion.create",
+            {
+                "title": "test_Xcdfgee",
+                "meeting_id": 222,
+                "workflow_id": 12,
+                "text": "test",
+                "lead_motion_id": 1,
+                "category_id": 12,
+                "block_id": 13,
+            },
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("motion/2", {"block_id": 13, "category_id": 12})
+
+    def test_create_amendment_no_perms_category_id(self) -> None:
+        self.create_meeting(222)
+        self.user_id = self.create_user("user")
+        self.login(self.user_id)
+        self.set_user_groups(self.user_id, [224])
+        self.set_group_permissions(
+            224,
+            [Permissions.Motion.CAN_CREATE, Permissions.Motion.CAN_CREATE_AMENDMENTS],
+        )
+        response = self.request(
+            "motion.create",
+            {
+                "title": "test_Xcdfgee",
+                "meeting_id": 222,
+                "workflow_id": 12,
+                "text": "test",
+                "lead_motion_id": 1,
+                "category_id": 12,
+            },
+        )
+        self.assert_status_code(response, 403)
+        assert "Forbidden fields: category_id" in response.json["message"]
+
+    def test_create_amendment_no_perms_block_id(self) -> None:
+        self.create_meeting(222)
+        self.user_id = self.create_user("user")
+        self.login(self.user_id)
+        self.set_user_groups(self.user_id, [224])
+        self.set_group_permissions(
+            224,
+            [Permissions.Motion.CAN_CREATE, Permissions.Motion.CAN_CREATE_AMENDMENTS],
+        )
+        response = self.request(
+            "motion.create",
+            {
+                "title": "test_Xcdfgee",
+                "meeting_id": 222,
+                "workflow_id": 12,
+                "text": "test",
+                "lead_motion_id": 1,
+                "block_id": 13,
+            },
+        )
+        self.assert_status_code(response, 403)
+        assert "Forbidden fields: block_id" in response.json["message"]
